Bugfixes : Gamma correction, stack allocation, number of threads

+++ sampling.h
@@ -3,6 +3,7 @@
 #include "geometry.h"
 #include "pcg/pcg_basic.h"
 #include <stddef.h> // size_t
+#include <stdint.h>
 
 #define STRATIFIED_RESOLUTION 4
 
@@ -15,5 +16,5 @@ void sample_unit_square(Sampler* sampler, float* x, float* y, float* pdf);
 void sample_unit_disc(Sampler* sampler, float* x, float* y, float* pdf);
 Vec3 sample_uniform_hemisphere(Sampler* sampler, float* pdf);
 Vec3 sample_cosine_weighted_hemisphere(Sampler* sampler, float* pdf);
-uint64_t sample_int(Sampler* sampler);
+uint32_t sample_int(Sampler* sampler);
 float rnd(Sampler* sampler);
